pyproject.toml
[project]
name = "waysklad"
version = "0.1.0"
description = ""
authors = [
    {name = "Rasul Yusupov",email = "raasaal511@gmail.com"}
]
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.135.2,<0.136.0)",
    "uvicorn (>=0.42.0,<0.43.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "greenlet (>=3.3.2,<4.0.0)",
    "black (>=26.3.1,<27.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
